refactor new chat modal into separate components and create forward menu
