+++ src/screens/Chat/ChatView.tsx
@@ -60,7 +60,7 @@ const ChatView: React.FC<Props> = ({ route, navigation }) => {
 
   const insets = useSafeAreaInsets();
   const FLATLIST_BOTTOM_OFFSET = 58 + insets.bottom;
-  const chatId = route.params.id;
+  const { id: chatId, forward } = route.params;
 
   useEffect(() => {
     if (!chatId || !chats || !chats.length) {
@@ -76,8 +76,18 @@ const ChatView: React.FC<Props> = ({ route, navigation }) => {
     setMessages(toViewable(chat.messages).reverse());
   }, [chats, chatId]);
 
+  useEffect(() => {
+    if (forward) {
+      addMessage({
+        chatId,
+        message: forward.text,
+        forwardedFrom: forward.from,
+      });
+    }
+  }, [route]);
+
   const pushMessage = (message: string) => {
-    addMessage(chatId, message);
+    addMessage({ chatId, message });
     scrollToBottom();
   };
